+++ validationframework-core/src/main/java/com/google/code/validationframework/base/property/simple/SimpleListProperty.java
@@ -253,6 +253,7 @@ public class SimpleListProperty<T> extends AbstractReadableWritableListProperty<
     }
 
     /**
+     * @see AbstractReadableWritableListProperty#contains(Object)
      * @see List#contains(Object)
      */
     @Override